+++ config.c
@@ -970,6 +970,12 @@ parse_section_colors(struct context *ctx)
     else if (strcmp(key, "text") == 0)
         return value_to_color(ctx, true, &conf->colors.text);
 
+    else if (strcmp(key, "prompt") == 0)
+        return value_to_color(ctx, true, &conf->colors.prompt);
+
+    else if (strcmp(key, "input") == 0)
+        return value_to_color(ctx, true, &conf->colors.input);
+
     else if (strcmp(key, "match") == 0)
         return value_to_color(ctx, true, &conf->colors.match);
 
@@ -1561,6 +1567,8 @@ config_load(struct config *conf, const char *conf_path,
             .background = conf_hex_to_rgba(0xfdf6e3dd),
             .border = conf_hex_to_rgba(0x002b36ff),
             .text = conf_hex_to_rgba(0x657b83ff),
+            .prompt = conf_hex_to_rgba(0x586e75ff),
+            .input = conf_hex_to_rgba(0x657b83ff),
             .match = conf_hex_to_rgba(0xcb4b16ff),
             .selection = conf_hex_to_rgba(0xeee8d5dd),
             .selection_text = conf_hex_to_rgba(0x657b83ff),